fix: Moving `awshelper` into an `internal` directory
